pyproject.toml
[project]
name = "aio-usb"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
license = { text = "MIT" }
authors = [
    { name = "David Lechner", email = "david@pybricks.com" }
]
requires-python = ">=3.11"
dependencies = [
    "rubicon-objc>=0.5.2 ; sys_platform == 'darwin'",
    "typing-extensions>=4.15.0",
    "winrt-windows-devices-enumeration>=3.2.1 ; sys_platform == 'win32'",
    "winrt-windows-devices-usb>=3.2.1 ; sys_platform == 'win32'",
    "winrt-windows-foundation>=3.2.1 ; sys_platform == 'win32'",
    "winrt-windows-foundation-collections>=3.2.1 ; sys_platform == 'win32'",
    "winrt-windows-storage>=3.2.1 ; sys_platform == 'win32'",
    "winrt-windows-storage-streams>=3.2.1 ; sys_platform == 'win32'",
]

[build-system]
requires = ["uv_build>=0.8.13,<0.9.0"]
build-backend = "uv_build"

[dependency-groups]
lint = [
    "ruff>=0.13.0",
]
